* src/include/font.h (font): Use `int' for ch_index. * src/libs/libgroff/font.cpp (font::alloc_ch_index, font::compact): Updated.
* src/roff/groff/pipeline.c (run_pipeline) [_WIN32]: Fix compiler
warnings.

* src/roff/troff/div.cpp (begin_page), src/roff/troff/env.cpp
(hyphen_word): Fix compiler warnings.
* src/roff/troff/input.cpp (get_char_for_escape_name): Return
`char'.
Update all callers.
(get_delim_number, get_line_arg): Change second argument type to
`unsigned char'.
(macro_header::copy, token::next, do_define_string,
do_define_character, substring_request, asciify_macro,
unformat_macro, read_size, non_interpreted_node::interpret,
while_request, main: Fix compiler warnings.
(read_color_draw_node): Initialize `col'.

* src/preproc/tbl/table.cpp (horizontal_span, table_entry): Make
`start_col' and `end_col' of type `int'.
(vertical_rule): Make `col' of type `int'.

* src/preproc/grn/hdb.cpp (DBRead): Protect call to getc with check
for EOF.

* src/preproc/refer/label.y (uppercase_array, lowercase_array):
New arrays.
(format_serial): Use them to remove dependency on ASCII.

* src/devices/grops/ps.cpp (ps_printer::define_encoding): Fix
compiler warning.
* src/devices/grops/psrm.cpp (resource_manager::supply_resource):
Ditto.

* src/devices/grotty/tty.cpp (tty_font::load_tty_font): Fix
compiler warnings.
(glyph): Change type of `hpos' to `int'.

* src/devices/lbp/lbp.cpp (strsep): Removed.  Unused.
* src/devices/lbp/lbp.h (splinerel): Fix compiler warnings.

* src/utils/tfmtodit.cpp (gf::load): Fix compiler warnings.

* src/utils/pfbtops/pfbtops.c (get_text): Fix compiler warning.

+++ src/libs/libgroff/font.cpp
@@ -403,7 +403,7 @@ void font::alloc_ch_index(int idx)
     nindices = 128;
     if (idx >= nindices)
       nindices = idx + 10;
-    ch_index = new short[nindices];
+    ch_index = new int[nindices];
     for (int i = 0; i < nindices; i++)
       ch_index[i] = -1;
   }
@@ -412,9 +412,9 @@ void font::alloc_ch_index(int idx)
     nindices *= 2;
     if (idx >= nindices)
       nindices = idx + 10;
-    short *old_ch_index = ch_index;
-    ch_index = new short[nindices];
-    memcpy(ch_index, old_ch_index, sizeof(short)*old_nindices);
+    int *old_ch_index = ch_index;
+    ch_index = new int[nindices];
+    memcpy(ch_index, old_ch_index, sizeof(int)*old_nindices);
     for (int i = old_nindices; i < nindices; i++)
       ch_index[i] = -1;
     a_delete old_ch_index;
@@ -443,9 +443,9 @@ void font::compact()
       break;
   i++;
   if (i < nindices) {
-    short *old_ch_index = ch_index;
-    ch_index = new short[i];
-    memcpy(ch_index, old_ch_index, i*sizeof(short));
+    int *old_ch_index = ch_index;
+    ch_index = new int[i];
+    memcpy(ch_index, old_ch_index, i*sizeof(int));
     a_delete old_ch_index;
     nindices = i;
   }
